# Conflicts:
#	usr/src/cmd/dfs.cmds/Makefile
#	usr/src/cmd/fs.d/nfs/mountd/Makefile
#	usr/src/cmd/smbsrv/smbadm/smbadm.c
#	usr/src/cmd/svc/svccfg/Makefile
#	usr/src/lib/Makefile
#	usr/src/lib/libc/port/print/doprnt.c
#	usr/src/pkg/manifests/service-network-network-clients.p5m
#	usr/src/tools/cw/cw.c
#	usr/src/tools/scripts/nightly
#	usr/src/uts/common/fs/zfs/zfs_vfsops.c
#	usr/src/uts/common/sys/Makefile.syshdrs

+++ usr/src/uts/common/exec/shbin/shbin.c
@@ -22,6 +22,7 @@
 /*
  * Copyright 2010 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
+ * Copyright 2019 Joyent, Inc.
  */
 
 #include <sys/types.h>
@@ -54,7 +55,7 @@ shbinexec(
 	struct uarg *args,
 	struct intpdata *idatap,
 	int level,
-	long *execsz,
+	size_t *execsz,
 	int setid,
 	caddr_t exec_file,
 	struct cred *cred,
@@ -158,7 +159,7 @@ shbinexec(
 	struct uarg *args,
 	struct intpdata *idatap,
 	int level,
-	long *execsz,
+	size_t *execsz,
 	int setid,
 	caddr_t exec_file,
 	struct cred *cred,
